requirements.txt
streamlit
pandas
python-dotenv
langchain-core
huggingface-hub
